History highlight integrated to parent collection view
