+++ Client/Frontend/Home/HistoryHighlights/FxHomeHistoryHightlightsViewModel.swift
@@ -8,6 +8,14 @@ import UIKit
 
 class FxHomeHistoryHightlightsViewModel {
 
+    struct UX {
+        static let maxNumberOfItemsPerColumn = 3
+        static let maxNumberOfColumns = 3
+        static let estimatedCellHeight: CGFloat = 65
+        static let verticalPadding: CGFloat = 8
+        static let horizontalPadding: CGFloat = 16
+    }
+
     // MARK: - Properties & Variables
     var historyItems: [HighlightItem]?
     private var profile: Profile
@@ -27,13 +35,13 @@ class FxHomeHistoryHightlightsViewModel {
     var numberOfColumns: Int {
         guard let count = historyItems?.count else { return 0 }
 
-        return Int(ceil(Double(count) / Double(HistoryHighlightsCollectionCellConstants.maxNumberOfItemsPerColumn)))
+        return Int(ceil(Double(count) / Double(UX.maxNumberOfItemsPerColumn)))
     }
 
     var numberOfRows: Int {
         guard let count = historyItems?.count else { return 0 }
 
-        return count < HistoryHighlightsCollectionCellConstants.maxNumberOfItemsPerColumn ? count : HistoryHighlightsCollectionCellConstants.maxNumberOfItemsPerColumn
+        return count < UX.maxNumberOfItemsPerColumn ? count : UX.maxNumberOfItemsPerColumn
     }
 
     /// Group weight used to create collection view compositional layout
@@ -120,15 +128,44 @@ extension FxHomeHistoryHightlightsViewModel: FXHomeViewModelProtocol, FeatureFla
                                  buttonA11yIdentifier: AccessibilityIdentifiers.FirefoxHomepage.MoreButtons.historyHighlights)
     }
 
-    // TODO: Laurie
-//    var isEnabled: Bool {
-//        guard featureFlags.isFeatureEnabled(.historyHighlights, checking: .buildAndUser) else { return false }
-//
-//        return !isPrivate
-//    }
-
     var isEnabled: Bool {
-        return false
+        guard featureFlags.isFeatureEnabled(.historyHighlights, checking: .buildAndUser) else { return false }
+
+        return !isPrivate
+    }
+
+    var section: NSCollectionLayoutSection {
+        let item = NSCollectionLayoutItem(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .estimated(UX.estimatedCellHeight))
+        )
+
+        let groupWidth = groupWidthWeight
+        let subItems = Array(repeating: item, count: numberOfRows)
+        let verticalGroup = NSCollectionLayoutGroup.vertical(
+            layoutSize: NSCollectionLayoutSize(widthDimension: groupWidth,
+                                               heightDimension: .estimated(UX.estimatedCellHeight)),
+            subitems: subItems)
+
+        let section = NSCollectionLayoutSection(group: verticalGroup)
+        section.orthogonalScrollingBehavior = .continuous
+        return section
+    }
+
+    var numberOfItemsInSection: Int {
+        guard let count = historyItems?.count else {
+            return 0
+        }
+
+        // If there are less than or equal items to the max number of items allowed per column,
+        // we can return the standard count, as we don't need to display filler cells.
+        // However, if there's more items, filler cells needs to be accounted for, so sections
+        // are always a multiple of the max number of items allowed per column.
+        if count <= UX.maxNumberOfItemsPerColumn {
+            return count
+        } else {
+            return numberOfColumns * UX.maxNumberOfItemsPerColumn
+        }
     }
 
     var hasData: Bool {
@@ -143,5 +180,191 @@ extension FxHomeHistoryHightlightsViewModel: FXHomeViewModelProtocol, FeatureFla
 
     func updatePrivacyConcernedSection(isPrivate: Bool) {
         self.isPrivate = isPrivate
+    }
+}
+
+// MARK: FxHomeSectionHandler
+extension FxHomeHistoryHightlightsViewModel: FxHomeSectionHandler {
+
+    func configure(_ cell: UICollectionViewCell,
+                   at indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = cell as? HistoryHighlightsCell else { return UICollectionViewCell() }
+
+        recordSectionHasShown()
+        
+        let hideBottomLine = isBottomCell(indexPath: indexPath,
+                                          totalItems: historyItems?.count)
+        let cornersToRound = determineCornerToRound(indexPath: indexPath,
+                                                    totalItems: historyItems?.count)
+        let shouldAddShadow = isBottomOfColumn(with: indexPath.row,
+                                               totalItems: historyItems?.count ?? 0)
+
+        guard let item = historyItems?[safe: indexPath.row] else {
+            return configureFillerCell(cell,
+                                       hideBottomLine: hideBottomLine,
+                                       cornersToRound: cornersToRound,
+                                       shouldAddShadow: shouldAddShadow)
+        }
+
+        if item.type == .item {
+            return configureIndividualHighlightCell(cell,
+                                                    hideBottomLine: hideBottomLine,
+                                                    cornersToRound: cornersToRound,
+                                                    shouldAddShadow: shouldAddShadow, item: item)
+        } else {
+            return configureGroupHighlightCell(cell,
+                                               hideBottomLine: hideBottomLine,
+                                               cornersToRound: cornersToRound,
+                                               shouldAddShadow: shouldAddShadow, item: item)
+        }
+    }
+
+    func didSelectItem(at indexPath: IndexPath,
+                       homePanelDelegate: HomePanelDelegate?,
+                       libraryPanelDelegate: LibraryPanelDelegate?) {
+
+        if let highlight = historyItems?[safe: indexPath.row] {
+            switchTo(highlight)
+        }
+    }
+
+    // MARK: - Cell helper functions
+
+    /// Determines whether or not, given a certain number of items, a cell's given index
+    /// path puts it at the bottom of its section, for any matrix.
+    ///
+    /// - Parameters:
+    ///   - indexPath: The given cell's `IndexPath`
+    ///   - totalItems: The number of total items
+    /// - Returns: A boolean describing whether or the cell is a bottom cell.
+    private func isBottomCell(indexPath: IndexPath, totalItems: Int?) -> Bool {
+        guard let totalItems = totalItems else { return false }
+
+        // First check if this is the last item in the list
+        if indexPath.row == totalItems - 1
+            || isBottomOfColumn(with: indexPath.row, totalItems: totalItems) { return true }
+
+        return false
+    }
+
+    private func isBottomOfColumn(with currentIndex: Int, totalItems: Int) -> Bool {
+        var bottomCellIndex: Int
+        for column in 1...numberOfColumns {
+            bottomCellIndex = (UX.maxNumberOfItemsPerColumn * column) - 1
+            if currentIndex == bottomCellIndex { return true }
+        }
+
+        return false
+    }
+
+    private func determineCornerToRound(indexPath: IndexPath, totalItems: Int?) -> UIRectCorner {
+        guard let totalItems = totalItems else { return [] }
+
+        var cornersToRound = UIRectCorner()
+
+        if isTopLeftCell(index: indexPath.row) { cornersToRound.insert(.topLeft) }
+        if isTopRightCell(index: indexPath.row, totalItems: totalItems) { cornersToRound.insert(.topRight) }
+        if isBottomLeftCell(index: indexPath.row, totalItems: totalItems) { cornersToRound.insert(.bottomLeft) }
+        if isBottomRightCell(index: indexPath.row, totalItems: totalItems) { cornersToRound.insert(.bottomRight) }
+
+        return cornersToRound
+    }
+
+    private func isTopLeftCell(index: Int) -> Bool {
+        return index == 0
+    }
+
+    private func isTopRightCell(index: Int, totalItems: Int) -> Bool {
+        let topRightIndex = (UX.maxNumberOfItemsPerColumn * (numberOfColumns - 1))
+        return index == topRightIndex
+    }
+
+    private func isBottomLeftCell(index: Int, totalItems: Int) -> Bool {
+        var bottomLeftIndex: Int {
+            if totalItems <= UX.maxNumberOfItemsPerColumn {
+                return totalItems - 1
+            } else {
+                return UX.maxNumberOfItemsPerColumn - 1
+            }
+        }
+
+        if index == bottomLeftIndex { return true }
+
+        return false
+    }
+
+    private func isBottomRightCell(index: Int, totalItems: Int) -> Bool {
+        var bottomRightIndex: Int {
+            if totalItems <= UX.maxNumberOfItemsPerColumn {
+                return totalItems - 1
+            } else {
+                return (UX.maxNumberOfItemsPerColumn * numberOfColumns) - 1
+            }
+        }
+
+        if index == bottomRightIndex { return true }
+
+        return false
+    }
+
+    private func configureIndividualHighlightCell(_ cell: UICollectionViewCell,
+                                                  hideBottomLine: Bool,
+                                                  cornersToRound: UIRectCorner,
+                                                  shouldAddShadow: Bool,
+                                                  item: HighlightItem) -> UICollectionViewCell {
+
+        guard let cell = cell as? HistoryHighlightsCell else { return UICollectionViewCell() }
+
+        let itemURL = item.siteUrl?.absoluteString ?? ""
+        let site = Site(url: itemURL, title: item.displayTitle)
+
+        let cellOptions = HistoryHighlightsViewModel(title: item.displayTitle,
+                                                     description: nil,
+                                                     shouldHideBottomLine: hideBottomLine,
+                                                     with: cornersToRound,
+                                                     shouldAddShadow: shouldAddShadow)
+
+        cell.updateCell(with: cellOptions)
+
+        getFavIcon(for: site) { image in
+            cell.heroImage.image = image
+        }
+
+        return cell
+    }
+
+    private func configureGroupHighlightCell(_ cell: UICollectionViewCell,
+                                             hideBottomLine: Bool,
+                                             cornersToRound: UIRectCorner,
+                                             shouldAddShadow: Bool,
+                                             item: HighlightItem) -> UICollectionViewCell {
+
+        guard let cell = cell as? HistoryHighlightsCell else { return UICollectionViewCell() }
+
+        let cellOptions = HistoryHighlightsViewModel(title: item.displayTitle,
+                                                     description: item.description,
+                                                     shouldHideBottomLine: hideBottomLine,
+                                                     with: cornersToRound,
+                                                     shouldAddShadow: shouldAddShadow)
+
+        cell.updateCell(with: cellOptions)
+
+        return cell
+
+    }
+
+    private func configureFillerCell(_ cell: UICollectionViewCell,
+                                     hideBottomLine: Bool,
+                                     cornersToRound: UIRectCorner,
+                                     shouldAddShadow: Bool) -> UICollectionViewCell {
+
+        guard let cell = cell as? HistoryHighlightsCell else { return UICollectionViewCell() }
+
+        let cellOptions = HistoryHighlightsViewModel(shouldHideBottomLine: hideBottomLine,
+                                                     with: cornersToRound,
+                                                     shouldAddShadow: shouldAddShadow)
+
+        cell.updateCell(with: cellOptions)
+        return cell
     }
 }
